Add files via upload
Hello, I'm stuck on these questions. I thought for the prediction it is asking us to use the test data but when I do this it comes up with an error message, I tried using a new dataset as well and this didn't work. 
Also I can't work out how to calculate RMSE and RSQ as the way I've used is what we have been taught. Please could I get some help on this?